Add unit tests for Input/Output using a separate shell script (#876)
The script starts instances of the PostgREST server in the background,
and runs requests with curl to check that the server accepts to verify
a JWT corresponding the configured secret.

The configuration file uses the trick value "@/dev/stdin" for jwt-secret,
ot allow reading a different secret from a relative path using the same
configuration file.

I also tried generating a dynamic configuration file but failed to
have PostgREST read it from standard input. I also tried to configure
relative paths in a configuration file, but failed to have PostgREST
to read the JWT secret from such a relative path.