Add binary secret with (LF) End Of Line (#877)
I added a final LF character to the file binary.noeol, which
does not have one:

  $ cp -p binary.noeol binary.eol
  $ echo '' >> binary.eol

The file now ends with LF character (0x0a, Line Feed):

  $ wc -c binary.noeol
         8 binary.noeol
  $ wc -c binary.eol
         9 binary.eol
  $ tail -c 1 binary.eol | xxd -p
  0a

Even for a file which contains a binary secret, any LF character found
at the end of the file shall be ignored. This is consistent with the
behavior of Unix utilities, e.g. `openssl dgst`:

  $ printf '%s' '{"role":"user"}' |
  > openssl dgst -sha256 -hmac "$(cat binary.noeol)" -hex
  7cc5980faa0d851e1cc41ad8dbfc7cc18662434991bdde8e3209277ed466a440

  $ printf '%s' '{"role":"user"}' |
  > openssl dgst -sha256 -hmac "$(cat binary.eol)" -hex
  7cc5980faa0d851e1cc41ad8dbfc7cc18662434991bdde8e3209277ed466a440

Both signatures are identical; the final LF character was not used
for the computation of the signature.