Add Base64 encoded version of each test secret (#877)
I encoded each secret in turn with the following command:

  for file in *.noeol
  do
    base64 -i "$file" -o "${file%.*}.b64"
  done

with the BSD base64 utility found on Mac OS X.

I used files without final End Of Line as input
because I found out that the final End Of Line is
not discarded in this case, but is encoded at the
end of the Base64 value.

Using the `openssl enc` utility instead:

  for file in *.noeol
  do
    openssl enc -base64 -in "$file" -out "${file%.*}.b64"
  done

produces identical files.