Add I/O tests for secrets encoded in Base64 (#877)
The configuration for Base64 secrets is selected based on the
extension of the secret (.b64) provided to the unit test utility
function readSecretFromFile. The configuration file itself is
identical but for the value of secret-is-base64 which is set to true.

I also added a test for the binary secret with final End Of Line.