Work around PostgreSQL 10 json_to_record change
This means that we keep accepting JSON embedded as strings
as arguments of type `json`/`jsonsb`.

The change explicitly converts potential json fields to string,
and only then to json in a second step, for PostgreSQL versions
from 10 on.

Accidentally, postgREST keeps accepting directly embedded JSON
fields with PostgreSQL >=10, via conversion to string and back
to JSON. This seems like a rather undesirable feature, but I'm
not clear how to avoid it.