Improve docker install examples
Removes the PGRST_DB_SCHEMA variable which is public by default now.

Removes the PGRST_DB_ANON_ROLE variable, because using the authenticator role or even a superuser is very bad practice.

Replaces the postgres superuser with app_user, because using a superuser to connect to postgres is very bad practice.

Signed-off-by: Wolfgang Walther <walther@technowledgy.de>