Update functions.rst
This change adds an explanation of how to handle an array of JSON objects within an RPC call.  To pass multiple objects, an array of JSON objects must be the JSON value, with the key being the json or jsonb variable name of the Postgres function.

For people who want to perform multiple tasks/inserts/updates within a single API call, this is a needed explanation for that use-case.

I recently had this problem, and after solving it on my own, I realized that I would have been saved several hours of troubleshooting had this simple example been included in the API documentation.