refactor: Remove all quasi quoters
While neat-interpolation builds fine in the static overlay, we still
can't actually *use* it for cross-compilation. Every quasi quotation
needs to execute Template Haskell during compilation, period.

Thus, get rid of it.

Currently, we keep heredoc for the test-suite. It would probably be a
**huge** pain to remove all quasi quotation from the spec tests, so we
don't even need to bother with that. As long as we can build the static
executable we should be good.